moved api calls to server side and cleaned up a lot of of stuff including old files
