+++ client/src/utils/random_recipe_code.js
@@ -1,4 +1,4 @@
-
+//this can probably go in react components directly because they'll be working with state
 
     function getRandomRecipe() {
         if (recipeSearchResults.length === 0) {
