+++ server/package.json
@@ -13,9 +13,12 @@
   "dependencies": {
     "@apollo/server": "^4.7.1",
     "bcrypt": "^5.1.1",
+    "crypto-js": "^4.2.0",
     "express": "^4.17.2",
     "graphql": "^16.6.0",
-    "mongoose": "^8.0.0"
+    "jssha": "^3.3.1",
+    "mongoose": "^8.0.0",
+    "oauth-1.0a": "^2.2.6"
   },
   "devDependencies": {
     "nodemon": "^2.0.6"
