+++ server/utils/UNUSED_food-ext-api/searchFoodsByKeywords.js
@@ -1,4 +1,3 @@
-import fetch from 'node-fetch';
 import OAuth from 'oauth-1.0a';
 import crypto from 'crypto';
